requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
mysqlclient==2.1.1
Pillow==9.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
